fix: Resolve all clippy warnings and errors
- Fix unused imports and variables (prefix with _ or remove)
- Fix dead code warnings (add config() method, allow for future use)
- Fix needless borrows and redundant closures
- Replace Error::new(ErrorKind::Other) with Error::other()
- Use unwrap_or_default() instead of unwrap_or_else(Default::default)
- Use derive Default instead of manual impls
- Fix port validation (remove impossible > 65535 checks for u16)
- Collapse nested if statements
- Fix enum variant names warning
- Remove ineffective .write(true) when .append(true) is used
- Remove unnecessary casts
- Fix let unit value and assertions on constants
- Add allow for useless_vec where Arrow requires Vec
- Add missing documentation for FileSpanExporter::new and FileMetricExporter::new
- Allow non_local_definitions for PyO3 bindings (known pyo3 macro limitation)
- Update CI workflow to include -A non_local_definitions flag for clippy
- Fix test files to use struct literal construction instead of Default + assignments

+++ src/otlp/server.rs
@@ -309,7 +309,7 @@ pub(crate) fn convert_trace_request_to_spans(
                     events: opentelemetry_sdk::trace::SpanEvents::default(),
                     links: opentelemetry_sdk::trace::SpanLinks::default(),
                     status,
-                    dropped_attributes_count: span.dropped_attributes_count as u32,
+                    dropped_attributes_count: span.dropped_attributes_count,
                     parent_span_is_remote: false,
                     instrumentation_scope,
                 };
